requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
apscheduler==3.10.4
pandas==2.1.3
jinja2==3.1.2
requests==2.31.0
alembic==1.12.1
aiosqlite==0.19.0
loguru==0.7.2
pydantic==2.5.0
pydantic-settings==2.1.0
celery==5.3.4
redis==5.0.1